Separated tests for latency
