+++ lib/latency.js
@@ -1,14 +1,4 @@
-'use strict';
-
-/**
- * Measure latency for a load test.
- * (C) 2013 Alex Fernández.
- */
-
-
-// requires
-const testing = require('testing');
-const crypto = require('crypto');
+import * as crypto from 'crypto'
 
 
 /**
@@ -18,7 +8,7 @@ const crypto = require('crypto');
  * An optional callback(error, results) will be called with an error,
  * or the results after max is reached.
  */
-class Latency {
+export class Latency {
 	constructor(options, callback) {
 		this.options = options
 		this.callback = callback
@@ -284,92 +274,5 @@ function createId() {
 	const value = '' + Date.now() + Math.random();
 	const hash = crypto.createHash('sha256');
 	return hash.update(value).digest('hex').toLowerCase();
-}
-
-/**
- * Test latency ids.
- */
-function testLatencyIds(callback) {
-	const latency = new Latency({});
-	const firstId = latency.start();
-	testing.assert(firstId, 'Invalid first latency id %s', firstId, callback);
-	const secondId = latency.start();
-	testing.assert(secondId, 'Invalid second latency id', callback);
-	testing.assert(firstId != secondId, 'Repeated latency ids', callback);
-	testing.success(callback);
-}
-
-/**
- * Test latency measurements.
- */
-function testLatencyRequests(callback) {
-	const options = {
-		maxRequests: 10,
-	};
-	const errorCode = '500';
-	const latency = new Latency(options, (error, result) => {
-		testing.check(error, 'Could not compute latency', callback);
-		testing.assertEquals(result.totalRequests, 10, 'Invalid total requests', callback);
-		testing.assertEquals(result.totalErrors, 1, 'Invalid total errors', callback);
-		testing.assert(errorCode in result.errorCodes, 'Error code not found', callback);
-		testing.assertEquals(result.errorCodes[errorCode], 1, 'Should have one ' + errorCode, callback);
-		testing.success(callback);
-	});
-	let id;
-	for (let i = 0; i < 9; i++) {
-		id = latency.start();
-		latency.end(id);
-	}
-	id = latency.start();
-	latency.end(id, errorCode);
-}
-
-/**
- * Check that percentiles are correct.
- */
-function testLatencyPercentiles(callback) {
-	const options = {
-		maxRequests: 10
-	};
-	const latency = new Latency(options, error => {
-		testing.check(error, 'Error while testing latency percentiles', callback);
-		const percentiles = latency.getResults().percentiles;
-
-		Object.keys(percentiles).forEach(percentile => {
-			testing.assert(percentiles[percentile] !== false, 'Empty percentile for %s', percentile, callback);
-		});
-
-		testing.success(percentiles, callback);
-	});
-	for (let ms = 1; ms <= 10; ms++) {
-		(function() {
-			const id = latency.start();
-			setTimeout(() => {
-				latency.end(id);
-			}, ms);
-		})();
-	}
-}
-
-/**
- * Run package tests.
- */
-function test(callback) {
-	const tests = [
-		testLatencyIds,
-		testLatencyRequests,
-		testLatencyPercentiles,
-	];
-	testing.run(tests, callback);
-}
-
-module.exports = {
-	Latency,
-	test,
-}
-
-// run tests if invoked directly
-if (__filename == process.argv[1]) {
-	test(testing.show);
 }
